dbのnumberをintegerからbigintに変更

+++ db/migrate/20230621122634_create_cows.rb
@@ -1,7 +1,7 @@
 class CreateCows < ActiveRecord::Migration[6.0]
   def change
     create_table :cows do |t|
-      t.integer :number,          null:false
+      t.bigint :number,          null:false
       t.date :birth_day
       t.text :memo
       t.references :user,         null: false,foreign_key: true
